chore(types): add shared/types.ts file and move List and ListItem types to it chore(types): add shared/* alias to babel.config.js and tsconfig.json files chore(main.test.ts): import ListItem and List types from shared/types.ts file instead of server/models/types.ts file chore(types.ts): remove ListItem and List types from dashboard/models/types.ts file and import them from shared/types.ts file instead
