+++ tsconfig.json
@@ -16,6 +16,7 @@
       "client/*": ["client/*"],
       "server/*": ["server/*"],
       "scheduler/*": ["scheduler/*"],
+      "shared/*": ["shared/*"],
       "@jest/*": ["./../.jest/*"],
       "@forms/*": ["./../docker/apply/forms-json/*"]
     },
